app/requirements.txt
fastapi==0.115.8
Jinja2==3.1.5
aiokafka==0.12.0
uvicorn==0.34.0
environs==10.3.0
python-dotenv==1.0.1
asyncpg==0.29.0
pypika==0.48.9